+++ bin/docker-magento.sh
@@ -12,6 +12,15 @@ Downloads the Docker Magento development environment from https://github.com/dan
 init () {
     git clone git@github.com:danhort/docker-magento.git ops
 
+    # Add path to git command
+    local git="git -C ./ops"
+
+    # Get the latest version
+    LATEST_VERSION=$(echo $($git tag) | awk '{print $NF}')
+
+    # Checkout the latest tag version
+    $git checkout -q $LATEST_VERSION
+
     # Copy the Makefile sample to the project root
     cp ops/samples/Makefile.sample ./Makefile
 
@@ -26,17 +35,37 @@ init () {
 }
 
 update () {
+    # Check if user root
     if [[ $EUID -ne 0 ]]; then
         echo "You have to be root to install Docker Magento. Try running with sudo."
         exit 1;
     fi
 
     echo 'Updating Docker Magento';
+
+    # Move to bin folder
     cd /usr/bin
-    wget https://raw.githubusercontent.com/danhort/docker-magento/master/bin/docker-magento.sh
+
+    local REMOTE_URL="https://github.com/danhort/docker-magento.git"
+
+    # Get the all the available tag versions
+    local AVIALABLE_VERSION=$(echo $(git ls-remote --refs -t -q $REMOTE_URL ) | tr " " "\n" | grep "refs" | tr "/" "\n" | grep "\.")
+
+    # Get the lates version
+    local LATEST_VERSION=$(echo $AVIALABLE_VERSION | awk '{print $NF}')
+
+    # Download the script
+    wget https://raw.githubusercontent.com/danhort/docker-magento/$LATEST_VERSION/bin/docker-magento.sh
+
+    # Reomove the old script
     rm docker-magento
+
+    # Rename the new script
     mv docker-magento.sh docker-magento
+
+    # Set correct file mode
     chmod +x docker-magento
+
     echo 'Update complete. Run Docker Magento with the command "docker-magento".'
     exit 0;
 }