+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The update command now checks for the latest tag
+- The magento-docker bash tool now initiates the latest tag
 
 ## [0.0.1] - 2020-04-29
 ### Added
@@ -13,4 +16,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/danhort/docker-magento/compare/0.0.1...HEAD
 [0.0.1]: https://github.com/danhort/docker-magento/releases/tag/0.0.1
-